reorganize the async parallel performer
